Placed the host in a separate package file
