+++ packages/e2e/host.js
@@ -1,0 +1,3 @@
+const host = "http://localhost:3000"; // @TODO read from environment variable
+
+module.exports = host;
