+++ packages/e2e/puppeteer/config.ts
@@ -1,9 +1,5 @@
 import { Flow } from "./types";
 
-// export const host = "https://vergunningcheck.amsterdam.nl/test";
-// export const host = "https://ux.chappie2.com";
-export const host = "http://localhost:3000";
-
 // @TODO: fix based on `CI`
 export const DEBUG = false;
